disable -Wdouble-promotion
It's more annoying than helpful.

+++ bench.c
@@ -80,7 +80,7 @@ int main(int argc, char** argv) {
 
     clock_t ticks = clock() - start;
     printf("%d loops in %g ticks, %.3g ns / pixel\n",
-            loops, (double)ticks, (double)ticks / (CLOCKS_PER_SEC * 1e-9) / (loops * NPIXELS));
+            loops, (double)ticks, ticks / (CLOCKS_PER_SEC * 1e-9) / (loops * NPIXELS));
 
     free(src_buf);
     free(dst_buf);
